Restructured project and updated structure after all merges for aesthetics
