+++ PythonPackageProject/requirements.txt
@@ -1,0 +1,13 @@
+Faker==30.8.1
+loguru==0.7.2
+numpy==2.1.2
+pandas==2.2.3
+psycopg2==2.9.10
+python-dateutil==2.9.0.post0
+python-dotenv==1.0.1
+pytz==2024.2
+six==1.16.0
+SQLAlchemy==2.0.36
+typing_extensions==4.12.2
+tzdata==2024.2
+streamlit==1.40.1
