--- PythonPackageProject/yourapplications/back/docker-compose.yml
@@ -1,31 +1,0 @@
-version: "3.8"
-
-services:
-  app:
-    build: .
-    container_name: fastapi_app
-    environment:
-      DATABASE_URL: postgresql://user:password@db/dbname
-    ports:
-      - "8000:8000"
-    depends_on:
-      - db
-    command: uvicorn back.main:app --host 0.0.0.0 --port 8000  # or adjust path to `app.main:app` if needed
-
-  db:
-    image: postgres:13
-    container_name: postgres_db
-    environment:
-      POSTGRES_USER: user
-      POSTGRES_PASSWORD: password
-      POSTGRES_DB: dbname
-    volumes:
-      - postgres_data:/var/lib/postgresql/data
-    ports:
-      - "5433:5432"
-
-volumes:
-  postgres_data:
-
-
-
