--- README.md
@@ -1,3 +1,0 @@
-# DS223-project
-
-Group 4
