--- mkdocs.yml
@@ -1,1 +1,0 @@
-site_name: My Docs
